feat(cli): add download single-mod command
The command will retrieve the details of the mod, find the selected
version, and download the file into the current directory. Types for the
relevant APIs have been added so that the functionality may be shared
with an upcoming feature that will accept a mod list and download all of
the mods.