feat: install mods from mod list file
This introduces a ron file where mod names and versions can be
specified. This is used to download all of the mods and create a
mod folder that is ready for the server to use. If the file a mod
intends to download already exists in the directory, then the download
will be skipped.