feat: launch factorio server using cli
The cli now supports a `start` command. This will accept a list of mods,
a directory to install them into, and will launch the game server after
setting up the mods. The game server runs as a child process with its
stdout and stderr copied into the parent process. Signal handlers for
SIGINT and SIGTERM have been added so that the server can be ended
gracefully.